requirements.txt
numpy~=1.19.5
pandas~=1.3.2
opencv-python~=4.5.3.56
deepface~=0.0.67
python-dotenv~=0.19.0
matplotlib~=3.4.2
os
glob
statistics~=1.0.3.5
imageio~=2.9.0
tensorflow~=2.5.0
Pillow~=8.3.1
scikit-image~=0.18.2
scipy~=1.7.1
tqdm~=4.62.0
py-feat